# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with Swagger documentation.",
        "title": "UnfallAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/unfallStrassenverkehrBundesland": {
            "get": {
                "description": "(46241-0022) Straßenverkehrsunfälle mit Personenschaden: Bundesländer, Jahre, Straßenklasse, Ortslage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Straßenverkehrsunfälle mit Personenschaden nach Bundesland abrufen",
                "parameters": [
                    {
                        "enum": [
                            "Baden-Württemberg",
                            "Bayern",
                            "Berlin",
                            "Brandenburg",
                            "Bremen",
                            "Hamburg",
                            "Hessen",
                            "Mecklenburg-Vorpommern",
                            "Niedersachsen",
                            "Nordrhein-Westfalen",
                            "Rheinland-Pfalz",
                            "Saarland",
                            "Sachsen",
                            "Sachsen-Anhalt",
                            "Schleswig-Holstein",
                            "Thüringen"
                        ],
                        "type": "string",
                        "description": "Filter nach Bundesland",
                        "name": "bundesland",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Autobahnen",
                            "Bundesstraßen",
                            "Landesstraßen",
                            "Kreisstraßen",
                            "Andere Straßen",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Straßenklasse",
                        "name": "strassenklasse",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "innerorts",
                            "außerorts",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Ortslage",
                        "name": "ortslage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter nach Jahr",
                        "name": "jahr",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UnfallStrassenverkehrBundesland"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameter type or range",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallStrassenverkehrBundesland/jahre": {
            "get": {
                "description": "Gibt alle Jahre zurück, für die Daten vorhanden sind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verfügbare Jahre abrufen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.YearsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallVerunglueckte": {
            "get": {
                "description": "(46241-0007/46241-0008) Verunglückte: Deutschland, Jahre, Monate, Geschlecht, Altersgruppen, Art der Verkehrsbeteiligung, Ortslage, Schwere der Verletzung",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verunglückte abrufen",
                "parameters": [
                    {
                        "enum": [
                            "Kraftrad mit Versicherungskennzeichen",
                            "Kraftrad mit amtlichem Kennzeichen",
                            "Elektrokleinstfahrzeuge",
                            "Personenkraftwagen",
                            "Kraftomnibus",
                            "Güterkraftfahrzeug",
                            "Landwirtschaftliche Zugmaschine",
                            "Übrige Kraftfahrzeuge",
                            "Fahrrad ohne Hilfsmotor",
                            "Pedelecs",
                            "Andere Fahrzeuge",
                            "Fußgänger",
                            "Andere Personen",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Art der Verkehrsbeteiligung",
                        "name": "verkehrsart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nach Ortslage (innerorts, außerorts (ohne Autobahnen), auf Autobahnen, Insgesamt)",
                        "name": "ortslage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Getötete",
                            "Schwerverletzte",
                            "Leichtverletzte",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Schweregrad",
                        "name": "schweregrad",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "männlich",
                            "weiblich",
                            "Ohne Angabe",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Geschlecht",
                        "name": "geschlecht",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "unter 15 Jahre",
                            "15 bis unter 18 Jahre",
                            "18 bis unter 21 Jahre",
                            "21 bis unter 25 Jahre",
                            "25 bis unter 35 Jahre",
                            "35 bis unter 45 Jahre",
                            "45 bis unter 55 Jahre",
                            "55 bis unter 65 Jahre",
                            "65 bis unter 75 Jahre",
                            "75 Jahre und mehr",
                            "Alter unbekannt",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Altersgruppe",
                        "name": "altersgruppe",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter nach Jahr",
                        "name": "jahr",
                        "in": "query"
                    },
                    {
                        "maximum": 12,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Filter nach Monat (1-12, 0 für Ganzjahresdaten)",
                        "name": "monat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UnfallVerunglueckte"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameter type or range",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallVerunglueckte/jahre": {
            "get": {
                "description": "Gibt alle Jahre zurück, für die Daten vorhanden sind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verfügbare Jahre abrufen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.YearsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.UnfallStrassenverkehrBundesland": {
            "type": "object",
            "properties": {
                "anzahl": {
                    "type": "integer"
                },
                "bundesland": {
                    "type": "string"
                },
                "jahr": {
                    "type": "integer"
                },
                "ortslage": {
                    "type": "string"
                },
                "strassenklasse": {
                    "type": "string"
                }
            }
        },
        "data.UnfallVerunglueckte": {
            "type": "object",
            "properties": {
                "altersgruppe": {
                    "type": "string"
                },
                "anzahl": {
                    "type": "integer"
                },
                "geschlecht": {
                    "type": "string"
                },
                "jahr": {
                    "type": "integer"
                },
                "monat": {
                    "description": "1-12 for months, 0 for full year data",
                    "type": "integer"
                },
                "ortslage": {
                    "type": "string"
                },
                "schweregrad": {
                    "type": "string"
                },
                "verkehrsart": {
                    "type": "string"
                }
            }
        },
...
        "handlers.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid parameter format"
                }
            }
        },
        "handlers.YearsResponse": {
            "type": "object",
            "properties": {
                "jahre": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        2022
                    ]
                }
            }
        }
    }
}